Rename vtt to core
Core of the application lives in there.
